SPIGOT-7555: Don't cast ItemFlags to byte
This fixes an issue with the HIDE_ARMOR_TRIM flag being lost during item
deserialization.

By: blablubbabc <lukas@wirsindwir.de>

+++ paper-server/src/main/java/org/bukkit/craftbukkit/inventory/CraftMetaItem.java
@@ -876,8 +876,8 @@ class CraftMetaItem implements ItemMeta, Damageable, Repairable, BlockDataMeta {
         return (this.hideFlag & bitModifier) == bitModifier;
     }
 
-    private byte getBitModifier(ItemFlag hideFlag) {
-        return (byte) (1 << hideFlag.ordinal());
+    private int getBitModifier(ItemFlag hideFlag) {
+        return 1 << hideFlag.ordinal();
     }
 
     @Override
